requirement.txt
SpeechRecognition
pyttsx3
PyAudio
wikipedia
simple-webbrowser
streamlit
pipwin
gTTS
google-generativeai
